Cadastro de funcionários/requirements.txt

fastapi
uvicorn[standard]
SQLAlchemy
jinja2
openpyxl
python-multipart
passlib[bcrypt]
python-jose